distsql/xeval: split eval.go into some small files (#1865)
* fix a bug when 'WHER' contains bit operationsE
* spilt distsql/xeval/eval.go into some small files
