+++ plan/expr_to_pb.go
@@ -122,17 +122,17 @@ func scalarFuncToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb
 	switch expr.FuncName.L {
 	case ast.LT, ast.LE, ast.EQ, ast.NE, ast.GE, ast.GT,
 		ast.NullEQ, ast.In, ast.Like:
-		return compareFuncToPBExpr(client, expr)
+		return compareOpsToPBExpr(client, expr)
 	case ast.Plus, ast.Minus, ast.Mul, ast.Div, ast.Mod, ast.IntDiv:
-		return arithmeticalFuncToPBExpr(client, expr)
+		return arithmeticalOpsToPBExpr(client, expr)
 	case ast.AndAnd, ast.OrOr, ast.UnaryNot:
-		return logicalFuncToPBExpr(client, expr)
+		return logicalOpsToPBExpr(client, expr)
 	default:
 		return nil
 	}
 }
 
-func compareFuncToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb.Expr {
+func compareOpsToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb.Expr {
 	var tp tipb.ExprType
 	switch expr.FuncName.L {
 	case ast.LT:
@@ -189,7 +189,7 @@ func compareFuncToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tip
 		Children: []*tipb.Expr{expr0, expr1}}
 }
 
-func arithmeticalFuncToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb.Expr {
+func arithmeticalOpsToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb.Expr {
 	var tp tipb.ExprType
 	switch expr.FuncName.L {
 	case ast.Plus:
@@ -221,7 +221,7 @@ func arithmeticalFuncToPBExpr(client kv.Client, expr *expression.ScalarFunction)
 		Children: []*tipb.Expr{expr0, expr1}}
 }
 
-func logicalFuncToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb.Expr {
+func logicalOpsToPBExpr(client kv.Client, expr *expression.ScalarFunction) *tipb.Expr {
 	var tp tipb.ExprType
 	switch expr.FuncName.L {
 	case ast.AndAnd:
